refactor(project): モジュール分割とインポート整理を実施
- main.py を最小構成にし、core/ 以下に機能別モジュールを分離
- 各インポートを PEP 8 に従って整理し、コメントを追加
- コードの可読性・保守性・拡張性を大幅に向上
